Add support for throwing HTTP errors with 5xx/6xx error codes which have a custom description
api-platform by default will replace the error message with a generic description of the
error code used.

In some cases we want to return server errors because some external service has
failed and we want to forward some information re why the request has failed.

This adds a new HTTPException subclass which gets its description set to the exception message.
The new class is added to avoid leaking internals for code using HTTPException and expecting
that the message isn't leaked for error codes >=500.

See #56