Keycloak: Handle tokens without a username set
For service accounts without any default scopes we get a null username from keycloak
through the token.

This handles that case by not accessing the username in case of a service account.

Also refactor the service account handling by just setting username to null for the
User layer if there is a service account (in theory the service account can have a dummy
username, but we can't use it for anything anyway because it is auto generated by keycloak)