Make the mapping between scopes and roles user definable
Instead of hardcoding the transformation let the user define
a method which translates the token scopes to smnfony roles.

This will allow us to remove the notion of scopes in the API
and let the user decide which scopes result in which roles,
and only roles are used for permission checks internally.