Add an API for attaching random values to a Person object
This can be used to forward data such as IDs for different services
to other bundles.

In the future we might replace this with a new API targeted at
translating user IDs between different formats, but this is good enough
for now to at least get rid of backend naming in the core bundle.