Change birthDate from DateTime to a string
Representing a date as a DateTime is hacky. Currently if you are
UTC-X then the date will be wrong. If we just force UTC then it's not
clear hwo the API client is supposed to handle that.

Also http://schema.org/Date specifies a Date and not a DateTime

Just use a string.