Remove all custom exceptions, except ApiError
We had this set of exceptions which were mapped to HTTP error codes
via api-platform. The problem is that there is no easy way to add them,
but we also have to repeat all existing mappings, which can easily break
if something with api-platform changes.

Also we want the bundle authors to think about which error code to return
and not use the easy to use exception which all map to the same error status.
Bundle authors should use HttpException or one of its subclasses, or ApiError().

All usage of these exceptions has been removed now from all bundles.

The only remaining exception is ApiError() which is special, in that it also
contains the original message even for 5xx codes, unlike HttpException.

See https://api-platform.com/docs/core/errors/