extension: use prependExtensionConfig for api_platform to allow overrides by the user
This for examples allows the user to enable doctrine support which otherwise wouldn't
be possible.

While this has the potential that the user overrides some defaults which we depend
on (and symfony doesn't merge values if there are multuple -> the last one wins)
we assume the user knows what he/she is doing.

If we decide in the future to officially support changing some options then we can
forward them from the core bundle config, like we currently do for the API docs
title and description.

Fixes #11