config: add messenger_transport_dsn and lock_dsn
These link the messenger and lock system for example to a redis server.

The lock compoennt will fall back to a local only backend like "semaphore"
when not specified.

The messenger will error out if a bundle requires it but there is no DSN set
(DSN means "data source name" btw)