Set up an in-memory transport when no messenger is needed and no queue is configured
In case the system admin has configured the queue commands via supervisor and something
in the app configuration changes so that the messenger is no longer used then suddenly
the queue commands would start to fail because no transport is configured.

Detect this case and install a dummy in-memory transport which does nothing. This
makes sure everything keeps working as before.

This also allows us to provide a supervisor configuration in the server template
without any actual messenger users in the server template.