queue: add support for doctrine transports
We set a default table name and group name with proper namespacing,
so it should be safe to reuse existing DBs.

This also adds a dbp:relay:core:queue:test command which queues
one or more dummy messages with an optional delay. This can be
used to see if the queue consumers work correctly and receive
tasks.