Only show the keycloak login button if keycloak is configured
In case the user doesn't provide all of the required parameters it will
fall back to the builtin swagger login button which requires the user
to paste the access token.

So in case a user doesn't use keycloak he/she can still use the API docs
with some valid access token.