Make debug mode an error during health checks
While it's mostly switched off when developing it being just a warning
makes it too easy to miss. Also we want to use the health check on
deployment and an enabled debug mode should make it fail there.

In the future we migth consider extra modes for less strict checks:

* Use more warnings but provide a "--strict" mode
* Provide a "--dev" mode which makes things that are OK during development not fail