Add a request specific ID to the logging processor
We currently attach a user session ID to each log entry, so we know
which requests are part of the same user session.

This only works after the user is authenticated since we generate the
logging ID from the access token.

This adds a per-request uuid so we know which log lines are part of
the same request.