Support multiple authenticators
In the future we want to support multiple authenticators depending on the request.
Since we want to define them in external bundles we register a proxy authenticator
which has a compiler pass and collects all other authenticators.

When a request comes in we go through all registered authenticators and forward the calls.

This also has the nice side effect that if no authenticators are registered we simple
fail to authenticate and don't error out when building the container because the
interface isn't implemented.

Other than that there should be no difference with our current setup.