Add a health check which initializes all services
In some cases where services can't be instantiated, either because the implementation
is missing, or some dependency is missing, or there is a cyclic dep then things would
fail only at the time the service is needed, for example in one specific request.

To avoid missing such errors and catch them early add a health check which tries to
build all services.

This allows us to abort a deployment in case a service is broken.