Add a ExtensionTrait which provides helpers for registering things in bundles
This trait can be used in bundle extension implementations to register various
paths and resources etc.

It hides various implementation details, so we can change how things are registered
in the future, and also makes it easier to discover what is possible and provide
proper documentation and type annotations directly in the code.

We can also do sanity checks and show deprecation warnings if needed.