Add a generic migration command: dbp:relay:core:migrate
With each bundle potentially having a DB and the doctrine migration
bundle requiring to explicitely pass the entity manager when running
a migration the admin has to know all entity manager names and
call them one by one to run all migrations.

Depending on the deployment environment and different bundle sets
this list can be different each time.

To make migrations easier and avoid missing one we register all
entity managers in the bundles, collect them and run all migrations
in one command.

While we are at it we also run the migrations for the worker queue,
which similarely creates DB tables for the queue messages.
This is especially important atm since auto setup is currently broken
in symfony and messenger:setup-transports needs to be called the first
time the API is used.

This doesn't take any options/arguments atm. We can add some once we have
a need. Ideally the bundles would configure what they need or tell us internally.