migrate: spawn separate processes for each migration
Turns out the migration bundle sets some global state which it
doesn't update on the next migration run, resulting in the wrong
entity manager being selected for some migrations.

Since it doesn't seem to be designed for calling it multiple times
in the same process try to isolate the calls by spawning a new
PHP process each time.

This assumes we find the right PHP executable and console command
on the filesystem, but hopefully this works in the common case.
If not we can try to use argv[0] for example.