Add a locale service for working with the the active locale
We currently configure symfony to apply the "Accept-Language"
header to all requests, so there remain only two cases where
we need to manually adjust things:

* In case an endpoint renders HTML that is displayed in an iframe
or popup we still need to provide an option to pass a language via
a query parameter. This then takes precedence over the header.

* In case we need a language identifier for another service, where
we need to request an external resource for a specific language.
In this case we either need to take the request language or in case
we are in a CLI context the default locale configured in symfony.