api-docs: extract the keycloak settings from the generic oidc config
In the future we plan to replace the keycloak specific web component used
in the API docs with a generic OIDC one. For this the auth bundle has now
started to set new global twig variables containing OIDC config and the keycloak
variabels are now deprecated.

In case we find the keycloak variables we will still use them, but if not
we will extract them from the OIDC url. This depends on the URL containing
a "/realms/" path element.

Once we move away from keycloak in the core we can remove this hack.