cron: add a "--force" option to the cron command
This will force run all cron commands even if they aren't due.
This can be helpful in case of development/debugging.

And we want to change the default of cron not running at the
beginning when the state/cache is missing i.e. when the api
is re-deployed. This allows the admin to force a run in case
it is known that the api was offline for some time and cron
jobs are needed.