cron: don't run all jobs if there was no previous run
In case the API is started the first time, or it is re-deployed
and the cache is cleared/removed the API doesn't know when the
cron jobs were last checked.

Up until now it would assume that the last run was so long ago that
all jobs are due and run all jobs. This has the downside that on
a re-deploy all jobs are run, despite most of them not needing to.

This changes the defaults to not run any jobs if there is no previous
check recorded.

If the admin still wants to force run all jobs there is the "--force"
option that can be passed to the cron command.