expression language: inject an object providing helper functions/methods
Instead of injecting global functions inject an object providing methods.
Mainly to have some namespacing in case the official language gains new
functions and to make it clear which function are from us and where the
documentation for them needs to be looked up.

The old global functions still remain for now to make the transition easier.