extension: add a way to register a new logging channel and configure masking
This allows a bundle to register an extension without having to create a
dummy service+tag.

Registering allows an optional parameter for disabling secret masking in the
log processor for that channel.

The main use case for this is when a bundle wants an additional audit channel
which can contain PII or similar.

Since we configure the monolog bundle directly here we need to add a dependency
on it and make sure it is loaded. This means that bundles might need to add/load
the monolog bundle in their tests now after updating to this version.