Add a dummy cache warmer which uses the application cache to setup the cache
In case the application cache uses a pdo adapter and there are no existing cache
saves it would spam the log with SQL errors due to cache misses, because it has
no logic to automatically create the database and table on a cache miss.

Work around this by adding a custom cache warmer which saves something to the
cache and deletes it again.