ExtensionTrait::addPathToHide: allow hiding operations other than GET
Up until now this only hid operations in the openapi docs for GET operations,
mainly because we usually want to hide the required GET item and collection
operations for resource where only POST does something useful and GET always
returns an empty collection or 404.

Another use case for hiding APIs is in case of renames, where we want to provide
the old name but not advertise it in the API docs.

To make this possible we now allow optionally passing an operation to
addPathToHide() to hide things besides GET.