composer: conflict with newer doctrine/dbal as a workaround
dependencies are currently split between supporting only 3 and some
also 4. Depending on the order other bundles are installed we might
get v4 first and due to composer being broken (it seems) it fails to
downgrade transitive dependencies again.

Add a conflicts so we stay on v3 for now. Remove once v4 is supported
by everything we depend on.