Add support for monolog v3
The main differences are that the ProcessorInterface now gets
a LogRecord instance instead of an array and also is expected
to return the same thing. And the level is now an object and not an int.

To support both versions we convert to an array, do our log processing
and convert back.

The tests are adjusted to just check the fields they care about instead of
the whole record, since the new class requires is more strict about which
fields need to exist.

Also add a test that checks if all fields are preserved, since we now convert
things between different types along the way.