Add support for api-platform 3.2 again
Turns out setting rfc_7807_compliant_errors gives us back most of
the old behaviour with more recent 3.2 versions, so do that.

Since json is disabled by default in api-platform 3.2, we need to enable
it on the controller or resource to test it. But ideally we want to not
change the defaults for all tests, so create a separate controller just
for that.