queue: implement a default failure transport
Currently the worker queue will by default retry a message three times
and then throw it away. This tries to use the Symfony failure transport
feature to make such cases easier to debug and the retry failed
messages at a later point.

To make this transparent to the user we derive everything based of
the queue_dsn config and set sensible defaults.

* We register a failure_transport for the default queue transport
  using the same DSN.
* We intercept the transport creation to properly configure the transport
  (queue_name, redis namespacing etc), so there are no conflicts
  with both using the same database connection.

With this, after a message was retried three times it will be re-added to
the failure transport and can be introspected via the Symfony CLI tools:

./bin/console messenger:failed:show --transport async_failed

Same for all the other "messenger:failed" subcommands.

To test this a new "--fail" option was added to the "dbp:relay:core:queue:test"
command. In addition various tests are added to improve the coverage. Since
introspecting a failure transport is currently only supported in Symfony for
doctrine transports we now depend and install a doctrine sqlite in-memory db
instead of the PHP in-memory transport. This requries more dev dependencies but
allows us to get more coverage.