symfony: backport new defaults to log PHP errors to the normal logging system
Currently (Symfony 5/6) it only logs PHP errors to monolog in debug mode.
Backport the default behaviour of Symfony 7 to do that always.

Currently PHP errors which would throw in debug mode only log a warning
in non-debug mode. Backport a fix from Symfony 7.1 to log an error in
those cases.