TestAuthenticator: provide a better error message when there is no user set.
This can happen in a unit test if a client is used for multiple requests,
since the symfony kernel will reboot by default, and the authenticator
will be added back, but no user will be set then.

It's not clear to me how to best improve this, so at least give a better
error message.

XXX: This uses AssertionError since that wont be converted to a 500 by
api-platform, otherwise the developer will only get 500 as an error
and not be shown the error message directly.