api docs login fixup
Turns out it doesn't add "Bearer" as the docs suggest.

It also doesn't unset when null is given. While setting null resets
the  swagger UI to show an empty input again, it sends an unknown
object instead of jsut not including the header.

So add "Bearer" to the value, and set an non-empty string, so it's clear
when debugging requests that there is no api key active.