Fix usage of TestUserAttributeProvider and don't just skip missing attributes
In c8363d25778fe5322554ce24200411ceb77ada0a the meaning of the constructor
parameter was changed from an array of attributes to a mapping of attributes
to default values, but the rest of the code wasn't adjusted.

This was missed because missing attributes which the provider advertised was
just ignored. Make this fatal, so this doesn't happen again.