Add new UserAttributeProviderExInterface which allows retrieving single attributes
UserAttributeProviderInterface has the downside that it only allows retrieving
all values at once. This usually isn't a problem, but we have started to
add conditions to the static provider which can itself retrieve other
attributes if defined in the config.

This can lead to fetching all attributes of one provider also fetching all
attributes of other providers (e.g. if the text file connector uses
an LDAP attribute in a condition) and this can lead to performance problems,
or lead to a downtime affecting more requests than needed.

Also add a hasUserAttribute() while at it. While unlikely that checking the
availability of all attributes is a performance issue, this allows us to optimize
things in the future if needed.

Connectors can switch to UserAttributeProviderExInterface if wanted.