UserSessionInterface: some cleanups and breaking changes
Rename getSessionTTL() to getSessionCacheTTL() to make it more clear. It was
always only used for cleaning up and never for enforcing any session validity,
so adjust the naming to match the cache key method.

Add a new isServiceAccount() which exposes if the user is a real user or just a
service account i.e. another machine. The naming is taken from what Keycloak uses
here.

Also in UserSession be more strict when calling methods when the user is not authenticated
and fail everywhere. Under normal usage that should never happen, but this makes sure
interface users don't depend on pre-auth session info. The only allowed call at any time is
to see if the user is authenticated.