ErrorProvider: fix debug mode not being forwarded to the service
It always defaulted to false, which means it removed the original exception
message also in debug mode.

Wire it up with the kernel.debug setting.

In the tests extend withUser() to allow setting kernel options, so we can
test with debug disabled, and add some tests that make sure that with debugging
disabled the trace is not serialized and the original exception message is not
serialized either.