ext: also set an in-memory queue transport in case there are unhandled messages
In case no transport is configured and there are messages that need routing still
set an in-memory transport. Otherwise in case anything needs the transport (
using the queue, or indirectly via doctrine) it would fail in the transport
factory.

By always setting one it will not fail, also not during bundle tests, and we
still have a health check which errors out in case a queue is needed but not
configured, which is much easier to understand for users compared to the transport
exception.