TestUtils: get rid of the TestUserSession service
Instead of overriding the UserSessionInterface with a test implementation
just use the default one. We already adjust everything needed for faking
a request via TestAuthenticator.

UserAuthWebTrait was still using it to fake a user, so port that to use
TestAuthenticator too, like the rest.

This means that the UserSessionInterface during unit tests will now behave
more like with a real request, for example being stricter with information
access in case the user is not authenticated.

TestUserSession can still be used as a dummy implementation in bundle tests,
but it's not part of the container any more.