openapi: add an Accept-Language header to each method
It's our default method for changing the language of a response and
adding it to every method is complicated, so use the decorator to
auto-add it to every method. Even if not every method makes use of/supports
different languages.

This makes it easier to test translations etc.

An alternative would have been to include our langauge selector at the
top for example, and intercept the openapi requests to add a header.
The problem there is that api-platform only allows passing static
data to the openapi js init, so we can't pass a function there.

If that changes in the future we could switch to that instead.